+++ MicroServices/MicroCuentas/Domain/Entities/Movimiento.cs
@@ -5,8 +5,8 @@
         public int id { get; set; }
         public DateTime fecha { get; set; }
         public string tipoMovimiento { get; set; }
-        public decimal valor { get; set; }
-        public decimal saldo { get; set; }
+        public double valor { get; set; }
+        public double saldo { get; set; }
         public int cuentaId { get; set; }
         public Cuenta? cuenta { get; set; }
     }
